gateway/requirements.txt
fastapi
uvicorn
requests
redis
paho-mqtt
python-dotenv